+++ server/pypi/packages/twisted/test.py
@@ -1,15 +1,13 @@
 from __future__ import absolute_import, division, print_function
 
-import sys
 import unittest
 
 
 class TestTwisted(unittest.TestCase):
 
     def test_basic(self):
-        # There's only one native module, and it's internal and Python 2-only, so an import
-        # test is sufficient.
-        from twisted.python import sendmsg
-        is_py2 = sys.version_info[0] < 3
-        self.assertEqual(is_py2, hasattr(sendmsg, "send1msg"))
-        self.assertEqual(is_py2, "twisted.python._sendmsg" in sys.modules)
+        # This is the only native module. It's for unit test purposes only, but the package
+        # build still requires it.
+        from twisted.test import raiser
+        with self.assertRaises(raiser.RaiserException):
+            raiser.raiseException()